requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
Werkzeug==2.3.6
zipp==3.15.0
celery==5.3.1
uWSGI==2.0.21
